separate GET ui config to no auth
